requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
aiohttp>=3.9.1
jinja2>=3.1.2
python-dotenv>=1.0.0
numpy>=1.24.3
pydantic>=2.5.0
requests
scikit-learn
python-dateutil
python-multipart

# Development and Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0